Updated link in Readme of hello-dapr-slim sample
Updated the instructions link under `Setup Dapr (Slim Init)` to point to the 'How-To: Run Dapr in self-hosted mode without Docker' article instead of the "How-To: Install Dapr into your local environment".  The self-hosted-no-docker article points to the original link as a prerequisite if the reader needs it. My fear was if they followed the directions on the original link it could be confusing because it requires docker and is not the slim init example.